// src/models/product.variation.option.model.ts
export interface ProductVariationOption {
  Id?: number;
  ProductVariationId: number;
  ProductId: string;
  VariationId: number;
  VariationOptionId: number;
  VariationName: string;
  Description: string;
  OptionName: string;
  CreateDate?: string;
  CreateUserId: string;
  ModifyDate?: string;
  ModifyUserId: string;
  StatusId: number;
}